refactor: 게시판 스크린 통합 및 category 기반 구조 리팩토링
